Rename environment.conf.json to app.json
Conditionally check if extended config exists.
