+++ .gitignore
@@ -6,7 +6,7 @@ stats.json
 .eslintcache
 
 # Local config
-environment.conf.json
+app.json
 
 # Cruft
 .DS_Store
