+++ README.md
@@ -26,7 +26,7 @@ The server listens on port 3001 by default. You can change the by setting the en
 
 ## Configuration
 
-Configuration for theming, map and API endpoint URLs is defined in `environment.conf.base.json`. Override the default configuration by creating the file `environment.conf.json`. This file is ignored by Git. Changes to the configuration file will only be picked up when the development server is restarted.
+Configuration for theming, map and API endpoint URLs is defined in `app.base.json`. Override the default configuration by creating the file `app.json`. This file is ignored by Git. Changes to the configuration file will only be picked up when the development server is restarted.
 
 ## Testing
 
@@ -64,7 +64,7 @@ Start the frontend server with a custom configuration as follows:
 
 ```bash
 docker run -d -p 8080:80 \
-  -v /branding/config.json:/environment.conf.json \
+  -v /branding/app.json:/app.json \
   -v /branding/logo.png:/usr/share/nginx/html/logo.png \
   -v /branding/favicon.png:/usr/share/nginx/html/favicon.png \
   -v /branding/manifest.json:/usr/share/nginx/html/manifest.json \
@@ -75,13 +75,13 @@ docker run -d -p 8080:80 \
 
 The `-v` flags mount a custom configuration from the host in the container:
 
-- `environment.conf.json` with specific configuration for theming, map and API endpoint URLs.
-- The logo, as defined by `logo.url` in `environment.base.conf.json`.
+- `app.json` with specific configuration for theming, map and API endpoint URLs.
+- The logo, as defined by `logo.url` in `app.base.json` (or `app.json`).
 - `favicon.png` the favicon, needs to be a PNG.
 - `manifest.json` with configuration for the PWA.
 - `icon_...x....png` the icons for the PWA in various sizes, need to be PNGs.
 
-The logo displayed on the website is defined by the `logo.url` parameter in `environment.base.conf.json`. This can be a remote URL or a local file path. Define for instance `/logo.svg` and you can inject your own `logo.svg` when starting the docker container.
+The logo displayed on the website is defined by the `logo.url` parameter in `app.base.json`. This can be a remote URL or a local file path. Define for instance (in `app.json`) `/logo.svg` and you can inject your own `logo.svg` when starting the docker container.
 
 The `manifest.json` file for the PWA is generated by webpack, but you can overwrite it with your own configuration. This file also defines some icons, which, in turn, can be injected into the container as well.
