+++ app.base.json
@@ -23,6 +23,7 @@
   },
   "language": {
     "avgDisclaimer": "De gemeente Amsterdam verwerkt uw persoonsgegevens op een zorgvuldige en veilige manier in overeenstemming met de geldende wet- en regelgeving, waaronder de Algemene Verordening Gegevensbescherming. Lees meer hierover op <a href='##AVG_DISCLAIMER_URL##'>Privacyverklaring Meldingen openbare ruimte en overlast</a>.",
+    "district": "Stadsdeel",
     "consentToContactSharing": "Ja, ik geef de gemeenten Amsterdam en Weesp toestemming om mijn contactgegevens te delen met andere organisaties, als dat nodig is om mijn melding goed op te lossen.",
     "footer1": "Bel het Gemeentelijk informatienummer: <a href='tel:##GENERAL_PHONE_NUMBER_FORMATTED##'>##GENERAL_PHONE_NUMBER##</a>",
     "footer2": "op werkdagen van 08.00 tot 18.00 uur.",
@@ -39,7 +40,24 @@
     "home": "https://www.amsterdam.nl",
     "privacy": "https://www.amsterdam.nl/privacy/specifieke/privacyverklaringen-wonen/meldingen-overlast-privacy/"
   },
-  "theme": {},
+  "theme": {
+    "primary": {
+      "main": "#004699",
+      "dark": "#00387a"
+    },
+    "secondary": {
+      "main": "#ec0000",
+      "dark": "#bc0000"
+    },
+    "support": {
+      "valid": "#00A03C",
+      "invalid": "#EC0000",
+      "focus": "#FEC813"
+    },
+    "error": {
+      "main": "#ec0000"
+    }
+  },
   "map": {
     "municipality": "amsterdam",
     "options": {
@@ -87,7 +105,10 @@
       "verlichting": "https://map.data.amsterdam.nl/maps/openbare_verlichting?REQUEST=GetFeature&SERVICE=wfs&OUTPUTFORMAT=application/json;%20subtype=geojson;%20charset=utf-8&Typename=Verlichting&version=1.1.0&srsname=urn:ogc:def:crs:EPSG::4326"
     }
   },
+  "areaTypeCodeForDistrict": "stadsdeel",
+  "fetchDistrictsFromBackend": false,
   "fetchQuestionsFromBackend": false,
+  "fetchSourcesFromBackend": false,
   "showVulaanControls": true,
   "useStaticMapServer": true,
   "apiBaseUrl": "https://acc.api.data.amsterdam.nl",
