+++ cypress/support/commandsCreateSignal.js
@@ -1,4 +1,4 @@
-import environmentConfig from '../../environment.conf.json';
+import appConfig from '../../app.base.json';
 import { CREATE_SIGNAL } from './selectorsCreateSignal';
 import { SIGNAL_DETAILS } from './selectorsSignalDetails';
 import { MANAGE_SIGNALS } from './selectorsManageIncidents';
@@ -35,23 +35,21 @@ export const checkDescriptionPage = () => {
 export const checkFlashingYellow = () => {
   if (Cypress.browser.name === 'firefox') {
     cy.log('No check on flashing yellow in Firefox');
-  }
-  else {
-    cy.get('.animate')
-      .then($selectors => {
-        const win = $selectors[0].ownerDocument.defaultView;
-        const after = win.getComputedStyle($selectors[0], 'after');
-        const contentValue = after.getPropertyValue('background-color');
-        expect(contentValue).to.eq('rgb(254, 200, 19)');
-      });
+  } else {
+    cy.get('.animate').then($selectors => {
+      const win = $selectors[0].ownerDocument.defaultView;
+      const after = win.getComputedStyle($selectors[0], 'after');
+      const contentValue = after.getPropertyValue('background-color');
+      expect(contentValue).to.eq('rgb(254, 200, 19)');
+    });
   }
 };
 
 export const checkHeaderFooter = () => {
-  cy.get(CREATE_SIGNAL.logoAmsterdam).should('have.attr', 'href', `${environmentConfig.links.home}`).and('be.visible');
+  cy.get(CREATE_SIGNAL.logoAmsterdam).should('have.attr', 'href', `${appConfig.links.home}`).and('be.visible');
   cy.get(`${CREATE_SIGNAL.disclaimer} h2`).should('have.text', 'Lukt het niet om een melding te doen?');
-  cy.contains(environmentConfig.language.footer2);
-  cy.get(CREATE_SIGNAL.footerPrivacyLink).should('have.attr', 'href', `${environmentConfig.links.privacy}`).and('be.visible');
+  cy.contains(appConfig.language.footer2);
+  cy.get(CREATE_SIGNAL.footerPrivacyLink).should('have.attr', 'href', `${appConfig.links.privacy}`).and('be.visible');
 };
 
 export const checkRedTextStatus = status => {
